pyproject.toml
[project]
name = "time_report"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "magnuswenzer", email = "magnus@wenzer.se"},
]
dependencies = [
    "sqlmodel>=0.0.22",
    "flet[all]>=0.24.1",
    "flet-contrib>=2024.3.6",
    "isoweek>=1.3.3",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "lxml>=6.0.2",
    "html5lib>=1.1",
    "beautifulsoup4>=4.14.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "python<3.14,>=3.8",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
